Backend/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5
Flask-CORS>=3.0.9
Flask-Testing>=0.8.1
flask-session==0.4.0
psycopg2-binary>=2.9
python-dotenv>=0.19.0

